### 2019-07-28 #1602 @whedon
[PRE REVIEW]: AutoGAMESS: A Python package for automation of GAMESS(us) Raman calculations
**Submitting author:** @Cavenfish (<a href="http://orcid.org/0000-0002-7416-8629">Brian Ferrari</a>)
**Repository:** <a href="https://github.com/Cavenfish/autogamess" target ="_blank">https://github.com/Cavenfish/autogamess</a>
**Version:** 1.0.31
