### 2019-07-28 #1602 @kyleniemeyer
Hi @Cavenfish, thanks for your interest in JOSS. Before we move forward with your submission, I have a few concerns based on our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) and [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html):

The biggest issue is that, while useful, the software appears (to me) to be more of a utility for interacting with another software package, AutoGAMESS, rather than fully fledged research software in its own right.  One common test for this is: if used towards a research paper, would it be cited? (Due to contributing meaningfully to the results.)

In addition, your submission is missing some of the major things we look for:
- I don't see any test suite, making it impossible for reviewers to assess whether the software works correctly.
- I don't see any documentation, or examples.
- I don't see any sort of community guidelines.
