### 2019-07-28 #1602 @Cavenfish
Hello Dr. Niemeyer,

This is my first submission to JOSS, so I am not sure how to continue from here. I understand rebuttals are typical in other journals, I’m not sure if this is the same in JOSS.

To address your concerns:

-AutoGAMESS is in fact a utility for interacting with another software package, however since it expedites the process of completing computational chemistry research it would certainly be cited, most likely within the acknowledgements section. It has already been acknowledged at conferences for aiding in the research progress. I’m also in contact with other researchers that plan on citing AutoGAMESS, once able to, In their future work.

-I will make and commit the test suit as soon as possible.

-The current level of documentation and guidelines is the doc strings for each function. If this is not sufficient I can write up documentation and community guidelines in the GitHub markdown as soon as possible.

Thank you for you time.

Thanks,
Brian Ferrari
