### 2019-07-28 7a27722
Fixed author and affilitation
