+++ Paper/paper.md
@@ -6,12 +6,11 @@ tags:
   - automation
   - raman
   - molecular properties
-author: |
- | Brian C. Ferrari
- | Department of Physics, University of Central Florida, 4111 Libra Drive, Orlando FL 32816
+author: Brian C. Ferrari
+affiliation: Department of Physics, University of Central Florida, 4111 Libra Drive, Orlando FL 32816
 orcid: 0000-0002-7416-8629
 geometry: margin=1in
-date: 27 May 2019
+date: 28 July 2019
 bibliography: paper.bib
 ---
 
@@ -30,7 +29,7 @@ an optimization and hessian calculation be performed on the molecule, making aut
 implemented individually by each research group utilizing the GAMESS(us) software. This slows scientific progress down, and an automation software written in a language extremely simple and well adopted by scientists, such as Python, is an attractive solution to the problem.
 AutoGAMESS is providing an open-source Python based software for automating conversion between Optimization calculations to
 Hessian calculations and then to Raman calculations. It also offers automation of data collection from the output files,
-for quick tabular data read outs of each calculation. AutoGAMESS has currently been used for a study presented at the 30th Annual Conference on Computation Physics CCP2018, its proceedings is soon to be published. AutoGAMESS is also currently being used in multiple other computational chemistry projects, soon to be published by scientists at the University of Central Florida. 
+for quick tabular data read outs of each calculation. AutoGAMESS has currently been used for a study presented at the 30th Annual Conference on Computation Physics CCP2018, its proceedings is soon to be published. AutoGAMESS is also currently being used in multiple other computational chemistry projects, soon to be published by scientists at the University of Central Florida.
 
 Functions:
 
