### 2019-07-29 #1602 @kyleniemeyer
@Cavenfish with JOSS all interactions with editors and reviewers are more like conversations rather than the more formal step-by-step process of older journals. In this case, the Associate Editor-in-Chief on duty (me) does a quick filter of new submissions before assigning to an editor. We look to ensure submissions are in scope, and aren't missing any major items.

For documentation, software packages should come with enough information such that a new user could come across the package and begin using it. At minimum, this could be in a README, but frequently more detailed documentation (like that generated via Sphinx) also accompanies software. This also generally includes installation info and examples.
