### 2019-07-29 dafd953
spelling mistakes fixed
