+++ CONTRIBUTE.md
@@ -1,0 +1,3 @@
+# Contributing to AutoGAMESS
+
+All contributions, big or small, are greatly appreciated. 
