### 2019-07-30 9a43b52
Added function documentation section
