+++ README.md
@@ -8,6 +8,275 @@ AutoGAMESS can be installed using
 
 `pip install autogamess`
 
+# Function Documentations
+`new_project(maindir,csvfile,initial_coords_dict=None,title='Project_Name/', make_inputs=False)`
+
+```
+This function creates a new directory tree for a GAMESS project, also makes
+  a couple of text files for use with other functions.
+
+  Parameters
+  ----------
+  maindir: string
+      A directory string (including the final `/`) that points to the
+      directory that the project tree will be spawned in.
+  csvfile: string
+      A directory string (including the final `.csv`) that points to the
+      text file containing project information. Read module documentation
+      for csv file format.
+  initial_coords_dict: dictionary [Optional]
+      This should be a dictionary with the key being the specie and the
+      value being a list that of its inital coordinates.
+  title: string [Optional]
+      A directory string (including the final `/`) that will be used as
+      the head of project directory tree.
+  make_inputs: boolean True/False [Optional]
+      if True then new_project will call input_builder at the end.
+
+  Notes 1
+  ----------
+  If the molecules you wish to build are not already defined in the
+  general autogamess coordinate dictionary, then initial_coords_dict
+  must be passed.
+
+  To see the autogamess coordianate dictionary simply print out
+  >>> ag.dictionaries.molecule_dictionary
+
+  Returns
+  ----------
+  This function returns nothing
+
+  Notes 2
+  ----------
+  The format of the spawned directory tree is as follows:
+
+                                  maindir
+                                     |
+                                   title
+                      -------------------------------------
+                      |        |        |        |        |
+                    Codes    Inps     Logs  Batch_Files  Spreadsheets
+                      |        |        |                     |
+                ---------    Block    -----------      1 file per specie
+                |       |             |    |    |
+          Text_Files  Scripts      Fail   Pass  Sorted
+                                    |      |        |
+                                 -------  Block   1 directory per specie
+                                 |     |
+                           Unsolved   Solved
+
+
+  Sections in directory tree labled 'Block' are directory trees with the
+  following format:
+                    1 directory per run type
+                                |
+                    1 directory per specie
+
+  Examples
+  ----------
+  >>> import autogamess as ag
+  >>>
+  >>> csvfile = './input.csv'
+  >>> maindir = './'
+  >>> title   = 'Project Title/'
+  >>>
+  >>> ag.new_project(maindir, csvfile, title=title)
+  >>>
+```
+
+`input_builder(inputfile, save_dir, initial_coords_dict=None,proj_title=' Your Title Goes Here\n')`
+
+```
+This function builds optimization input files.
+
+Parameters
+----------
+inputfile: string
+    This should be a full directory string that points to the input
+    csv file.
+save_dir: string
+    this should be a full directory string that points to the directory
+    you wish to save the inputs in.
+initial_coords_dict: dictionary [Optional]
+    This should be a dictionary with the key being the specie and the
+    value being a list that of its inital coordinates.
+proj_title: string [Optional]
+    This should be a string ending with `\n`
+
+Notes 1
+----------
+If the molecules you wish to build are not already defined in the
+general autogamess coordinate dictionary, then initial_coords_dict
+must be passed.
+
+To see the autogamess coordianate dictionary simply print out
[…]
