### 2019-07-30 c042ee3
Added subheaders to function documentation section
