### 2019-07-30 23bbaa1
Wrote full Community Guidlines
