+++ CONTRIBUTE.md
@@ -1,3 +1,19 @@
 # Contributing to AutoGAMESS
 
-All contributions, big or small, are greatly appreciated. 
+All contributions, big or small, are greatly appreciated. If you wish to contribute some changes to the code then you should submit a pull request.
+
+### Pull Requests
+
+You can find detailed instructions and a discussion of how to fork a repository, clone it locally and work on the changes in the [GitHub guides](https://guides.github.com/activities/forking/).
+
+### Bugs and Feature Requests
+
+The most effective way to spread awareness of bugs or feature requests is by raising an issue on the github issue tracker. Github allows you to classify your issues so that we know if it is a bug report, feature request or feedback to the authors.
+
+# Tests
+
+AutoGAMESS uses pytest for testing, from within the tests directory execute the following to run tests.
+
+`python -m pytest`
+
+When contributing always confirm all tests complete successfully prior to making changes, if not maybe you can help find and fix the problem. Once changes are made all tests should pass and if new functionality was introduced tests for that should be made and passed.
