### 2019-07-30 044ec3e
got rid of unneeded code
