### 2019-07-30 d760d4d
added test for bond angle and length calculations
