+++ tests/test_package.py
@@ -2,6 +2,8 @@ import pytest
 import os
 import filecmp as fc
 import autogamess as ag
+import numpy as np
+from scipy.spatial import distance
 
 csvfile = './input.csv'
 title   = 'check/'
@@ -24,3 +26,10 @@ def test_input_builder(tmpdir):
         assert 0
 
     return
+
+def test_bond_length_and_angle():
+    a1 = ag.make_xzy([0,0,0])
+    a2 = ag.make_xzy([1,0,0])
+    assert distance.euclidean(a1, a2) == 1
+    a1 = ag.make_xzy([0,1,0])
+    assert ag.angle_between(a1, a2) == (np.pi/2)
