+++ autogamess/hes2raman.py
@@ -55,7 +55,7 @@ def hes2raman(hesfile, datfile):
     f.close()
 
     #checks if they are in file
-    if check_if_in(datfile, ctr_f(s,grab), ctr_f(e,grab)):
+    if check_if_exists(datfile, ctr_f(s,grab), ctr_f(e,grab)):
         raise ValueError()
 
     #Define start and end indexes
