### 2019-07-30 deac3c5
added get_data example
