### 2019-07-30 1b0e69b
Fixed input_builder test and finished get_data test
