+++ README.md
@@ -243,7 +243,7 @@ Examples
 
 ### fill_spreadsheets
 
-`fill_spreadsheets(projdir=False, sorteddir=False, sheetsdir=False)`
+**`fill_spreadsheets(projdir=False, sorteddir=False, sheetsdir=False)`**
 
 ```
 This function fills in the spreadsheets initially generated by new_project
