### 2019-07-30 1d7a47d
Bolded function call displays
