### 2019-07-30 #1602 @Cavenfish
@labarba I believe I have fully satisfied all the previously addressed concerns. I have added documentation and examples to the README.md and added community guidelines to a CONTRIBUTE.md.  A test suite was also added for the package.

Is this ready to proceed to the review stage, or should i make more updates first?

Thanks!
