### 2019-07-30 ed0f361
added sentance to use of AutoGAMESS section
