### 2019-07-30 #1602 @labarba
👋 @jedbrown — Could you handle this new JOSS submission as editor? Thanks!
