### 2019-07-30 #1602 @jedbrown
@arosen93 @goerz @shivupa @robertodr @KEIPERTK :wave: Would you be willing to review this submission for JOSS?
