### 2019-07-31 #1612 @jedbrown
@arosen93, @shivupa, and @colleeneb :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [AutoGAMESS repository](https://github.com/Cavenfish/autogamess)). I'll be watching this thread if you have any questions.
