### 2019-08-01 #1612 @jedbrown
@shivupa We can set the version at acceptance. It's common for the review to result in a new release.
