### 2019-08-01 Cavenfish/autogamess#1 @shivupa
Paper DOIs
For review [#1612](https://github.com/openjournals/joss-reviews/issues/1612): Could you update the `references.bib` to include DOIs for the papers.
