### 2019-08-02 0e934e1
added doi's to bitex refs fixes #1
