+++ Paper/paper.bib
@@ -4,6 +4,7 @@
   journal={Journal of computational chemistry},
   volume={14},
   number={11},
+  doi={10.1002/jcc.540141112},
   pages={1347--1363},
   year={1993},
   publisher={Wiley Online Library}
@@ -15,6 +16,7 @@
   journal={Journal of Molecular Structure: THEOCHEM},
   volume={683},
   number={1-3},
+  doi={10.1016/j.theochem.2004.06.013},
   pages={183--193},
   year={2004},
   publisher={Elsevier}
@@ -26,6 +28,7 @@
   journal={Journal of computational chemistry},
   volume={34},
   number={31},
+  doi={10.1002/jcc.23415},
   pages={2668--2676},
   year={2013},
   publisher={Wiley Online Library}
@@ -37,6 +40,7 @@
   journal={Astronomy \& Astrophysics},
   volume={438},
   number={1},
+  doi={10.1051/0004-6361:20052658},
   pages={31--37},
   year={2005},
   publisher={EDP Sciences}
@@ -48,6 +52,7 @@
   journal={The Astrophysical Journal},
   volume={782},
   number={2},
+  doi={10.1088/0004-637x/782/2/63},
   pages={63},
   year={2014},
   publisher={IOP Publishing}
@@ -59,6 +64,7 @@
   journal={Journal of computer-aided molecular design},
   volume={14},
   number={2},
+  doi={10.1023/A:1008193805436},
   pages={123--134},
   year={2000},
   publisher={Springer}
@@ -67,6 +73,7 @@
 @article{schmidt2013webmo,
   title={WebMO Enterprise, version 13.0},
   author={Schmidt, JR and Polik, WF},
+  url={http://www.webmo.net},
   journal={WebMO LLC},
   year={2013}
 }
@@ -77,6 +84,7 @@
   journal={Journal of Molecular Graphics and Modelling},
   volume={16},
   number={3},
+  doi={10.1016/S1093-3263(99)00002-9},
   pages={133--138},
   year={1998},
   publisher={Elsevier}
@@ -88,6 +96,7 @@
   journal={Journal of computational chemistry},
   volume={32},
   number={1},
+  doi={10.1002/jcc.21600},
   pages={174--182},
   year={2011},
   publisher={Wiley Online Library}
@@ -99,6 +108,7 @@
   journal={Trends in cell biology},
   volume={15},
   number={7},
+  doi={10.1016/j.tcb.2005.05.001},
   pages={356--363},
   year={2005},
   publisher={Elsevier}
@@ -110,6 +120,7 @@
   journal={Journal of computational chemistry},
   volume={29},
   number={5},
+  doi={10.1002/jcc.20823},
   pages={839--845},
   year={2008},
   publisher={Wiley Online Library}
@@ -121,6 +132,7 @@
   journal={Journal of computational chemistry},
   volume={17},
   number={13},
+  doi={10.1002/(SICI)1096-987X(199610)17:13<1571::AID-JCC9>3.0.CO;2-P},
   pages={1571--1586},
   year={1996},
   publisher={Wiley Online Library}
@@ -132,6 +144,7 @@
   journal={Journal of chemical information and modeling},
   volume={47},
   number={3},
+  doi={10.1021/ci600510j},
   pages={1045--1052},
   year={2007},
   publisher={ACS Publications}
