### 2019-08-02 #1612 @Cavenfish
An issue with AutoGAMESS angle calculations has recently been pointed out to me. I should have this fixed in less than a day, the changes will be reflected in the paper as well.
I apologize for not catching this sooner.
