### 2019-08-02 dd8cf2e
added function to translate origin for angle calculations
