### 2019-08-02 559766a
fixed issure with bond angle calculation
