### 2019-08-02 cc2b806
added test for new origin shifted angle calculation
