### 2019-08-02 2ef8fc0
revised angle calculation explanation after fix in code
