### 2019-08-02 #1612 @Cavenfish
Okay, I have committed the edits to the code and paper.
