### 2019-08-03 Cavenfish/autogamess#2 @Andrew-S-Rosen
Comments on the JOSS paper
This is part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1612). The following consists of several comments on the paper aspect of the JOSS submission and does not reflect the content of the underlying code.

With regards to the "Summary" section:

1. The author states that "... there still is no open-source software for automation of ab initio calculations." Over the past several years, there have been many efforts to develop open-source software that can automatically generate input files for high-throughput quantum-chemical calculations. There are too many examples to list here. However, a few examples include [[1](https://iopscience.iop.org/article/10.1088/1361-648X/aa680e)], [[2](https://www.sciencedirect.com/science/article/pii/S0927025617303919)], and [[3](https://github.com/jyosa/MaPhi)]. There are many others aimed at specific domains of research (e.g. [[4](https://pubs.acs.org/doi/abs/10.1021/acs.jpca.7b07361)]). I would therefore refrain from making this claim in the paper. If the author is confident that no other program exists with the specific purpose of automating Raman calculations in GAMESS, they can state this. Otherwise, the author should simply emphasize the importance of automated workflows in quantum-chemical applications as the motivation for this work.

2. "Often times single calculations of molecular properties are not reliable resulting is publications requiring several calculations for each property be done on each molecule". It is not clear what exactly this sentence means. Why is a single calculation not reliable, and what is being changed with the subsequent calculations? Does the author mean that several levels of theory must be used or multiple input parameters must be considered?

3. The author motivates the work by stating that GAMESS calculations are done to predict properties of volatile and unstable species that have not been experimentally characterized or quantized before. Certainly there will be users of this code that will compute properties of known molecules. It is not restricted to molecules that have not been characterized.

With regards to the "Capabilities" section:

1. As specified by the BSE webpage (and corresponding pypi package), the author should cite "New Basis Set Exchange: An Open, Up-to-date Resource for the Molecular Sciences Community. Benjamin P. Pritchard, Doaa Altarawy, Brett Didier, Tara D. Gibson, Theresa L. Windus. 2019, Manuscript in Preparation" to refer to the most up-to-date version of the EMSL BSE.

2. "while optimization properties are calculated". It is not clear what "optimization properties" means. Properties computed from the geometry optimization? Please clarify this sentence.

With regards to the "Availability" section:

1. Please correct the install command to read `python -m pip install autogamess --user` (`python` is the more widely used command to execute Python codes).

2. Could the author please include the link to the [pypi page](https://pypi.org/project/autogamess/) for the code so the user knows where `pip` is pulling the data from?

With regards to the "Dependencies" section:

1. If there are minimum versions for the dependencies that are required for AutoGAMESS to work, please specify them.

General comments:

1. The `setup.py` file specifies the user should use Python 3.x. However, this is not mentioned in the repository or paper. The required Python version should be specified.

2. The paper has many grammatical errors that need to be corrected. A non-exhaustive list is included below, along with some needed clarifications:
- In general, "software" cannot be made plural. "Softwares" should be rephrased appropriately (for instance, to "programs" or "codes" or "packages" or "tools" or similar).
- Line 5: the comma should be removed.
- Line 6: "Making GAMESS(us)..." is not a complete sentence
- "However, these..." should read "However, because these"
- "Computational Chemistry" should not be capitalized
- "with multiple steps for each final result" is not clear wording. It should be rephrased to something like "with multiple steps required to achieve each final result"
- "for instance a raman" should be the start of a new sentence
- All instances of "raman" should be written as "Raman"
- "first requires an optimization". Please specify that this is a geometry optimization
- "often times" should read "oftentimes"
- "resulting is publications" should read "resulting in publications"
- A comma should be added after "As it stands"
- "AutoGAMESS is providing" should read "AutoGAMESS provides"
- All instances of "hessian" should be made "Hessian"
- "Optimization calculations" should be made lowercase
- "read outs" should be written as "readouts"
- "This main directory will also generate spreadsheets" is not grammatically correct
- All instances of "ESML" should read "EMSL"
- A comma should be included between "optimization calculation" and "AutoGAMESS is able"
- A comma should be included after "have been completed"
- "cartesian" should be made "Cartesian"
- "is the coordinates" should be made "are the coordinates"
- "in its usability, however" should read "in its usability; however" or "in its usability. However,"
- Please put the appropriate parenthesis in the arccos function.
- Please specify that the V\hat variables are normalized _position_ vectors. Also, please specify if it is the user-specified origin or the modified origin (i.e. the third atom).
- Please capitalize "python" to "Python"
- "softwares GitHub" should read "software's GitHub"
- Please define _A_ as the angle.
