### 2019-08-04 3017db2
Made revisions to paper fixes #2
