+++ Paper/paper.md
@@ -4,7 +4,7 @@ tags:
   - Python
   - computational chemistry
   - automation
-  - raman
+  - Raman
   - molecular properties
 authors:
   - name: Brian C. Ferrari
@@ -21,53 +21,45 @@ bibliography: paper.bib
 # Summary
 
 The *ab initio* Quantum Chemistry software GAMESS(us)[@schmidt1993general] is capable of calculating a variety of molecular
-properties. The software is widely used by chemists, physicists, astro-chemists and astro-physicists
-[@burda2004theoretical; @pacifici2013high; @hickman2005dissociative; @bennett2014experimental], for predicting properties of
-volatile and unstable species that have not been experimentally characterized or quantized before. Making GAMESS(us) of high
-importance to these research groups, because of this large importance numerous third-party softwares have been written to compliment GAMESS(us)
-[@schaftenaar2000molden; @schmidt2013webmo; @bode1998macmolplt; @allouche2011gabedit; @dermardirossian2005gdis]. However,
-these softwares are largely visualization and graphical softwares, there still is no open-source software for automation
-of *ab initio* calculations. Softwares for data collection from the output files of Computation Chemistry softwares have been made[@o2008cclib], but automation of input generation is not included. Research utilizing
-*ab initio* calculations typically require calculations with multiple steps for each final result, for instance a raman activity prediction first requires
-an optimization and hessian calculation be performed on the molecule, making automation extremely beneficial. Often times single calculations of molecular properties are not reliable resulting is publications requiring several calculations for each property be done on each molecule. As it stands the automation of these calculations is either not being done, or being
-implemented individually by each research group utilizing the GAMESS(us) software. This slows scientific progress down, and an automation software written in a language extremely simple and well adopted by scientists, such as Python, is an attractive solution to the problem.
-AutoGAMESS is providing an open-source Python based software for automating conversion between Optimization calculations to
-Hessian calculations and then to Raman calculations. It also offers automation of data collection from the output files,
-for quick tabular data read outs of each calculation. AutoGAMESS has currently been used for a study presented at the 30th Annual Conference on Computation Physics CCP2018, its proceedings is soon to be published. AutoGAMESS is also currently being used in multiple other computational chemistry projects, soon to be published by scientists at the University of Central Florida.
+properties. One of the many popular uses of GAMESS(us) is by chemists, physicists, astro-chemists and astro-physicists
+[@burda2004theoretical; @pacifici2013high; @hickman2005dissociative; @bennett2014experimental] for predicting properties of
[…]
