+++ Paper/paper.bib
@@ -149,3 +149,43 @@
   year={2007},
   publisher={ACS Publications}
 }
+
+@article{EMSL,
+  title={A New Basis Set Exchange: An Open, Up-to-date Resource for the Molecular Sciences Community},
+  author={Benjamin P. Pritchard, Doaa Altarawy, Brett Didier, Tara D. Gibson, Theresa L. Windus},
+  year={2019},
+  journal={Manuscript in Preparation}
+}
+
+@article{Bhoorasingh,
+  author={Bhoorasingh, Pierre L. and Slakman, Belinda L. and Seyedzadeh Khanshan, Fariba and Cain, Jason Y. and West, Richard H.},
+  title={Automated Transition State Theory Calculations for High-Throughput Kinetics},
+  journal={The Journal of Physical Chemistry A},
+  volume={121},
+  number={37},
+  pages={6896-6904},
+  year={2017},
+  doi={10.1021/acs.jpca.7b07361}
+}
+
+@article{MATHEW2017140,
+  title={Atomate: A high-level interface to generate, execute, and analyze computational materials science workflows},
+  journal={Computational Materials Science},
+  volume={139},
+  pages={140 - 152},
+  year={2017},
+  doi={10.1016/j.commatsci.2017.07.030},
+  author={Kiran Mathew, Joseph H. Montoya, Alireza Faghaninia, Shyam Dwarakanath, Muratahan Aykol, Hanmei Tang, Iek-heng Chu, Tess Smidt, Brandon Bocklund, Matthew Horton, John Dagdelen, Brandon Wood, Zi-Kui Liu, Jeffrey Neaton, Shyue Ping Ong, Kristin Persson, Anubhav Jain}
+}
+
+@article{Hjorth_Larsen_2017,
+	doi={10.1088/1361-648x/aa680e},
+	year={2017},
+	publisher={{IOP} Publishing},
+	volume={29},
+	number={27},
+	pages={273002},
+	author={Ask Hjorth Larsen and Jens J{\o}rgen Mortensen and Jakob Blomqvist and Ivano E Castelli and Rune Christensen and Marcin Du{\l}ak and Jesper Friis and Michael N Groves and Bj{\o}rk Hammer and Cory Hargus and Eric D Hermes and Paul C Jennings and Peter Bjerre Jensen and James Kermode and John R Kitchin and Esben Leonhard Kolsbjerg and Joseph Kubal and Kristen Kaasbjerg and Steen Lysgaard and J{\'{o}}n Bergmann Maronsson and Tristan Maxson and Thomas Olsen and Lars Pastewka and Andrew Peterson and Carsten Rostgaard and Jakob Schi{\o}tz and Ole Schütt and Mikkel Strange and Kristian S Thygesen and Tejs Vegge and Lasse Vilhelmsen and Michael Walter and Zhenhua Zeng and Karsten W Jacobsen},
+	title={The atomic simulation environment{\textemdash}a Python library for working with atoms},
+	journal={Journal of Physics: Condensed Matter}
+}
