### 2019-08-04 #1612 @Cavenfish
@arosen93  Your comments on the paper were greatly appreciated! I have made revisions to the paper taking your comments into account, and I believe it has strengthened the paper.

I look forward to hearing more comments from you all.
