### 2019-08-05 Cavenfish/autogamess#3 @Andrew-S-Rosen
Installation error
This is part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1612).

The installation instructions say to use `pip` to install AutoGAMESS. However, it seems that this does not install the required dependencies. There are several ways to address this, but can it be made such that these dependencies are automatically installed?

In addition, please correct the `Periodic_Elements` package name in both the paper and README. It should be `PeriodicElements`.
