### 2019-08-05 Cavenfish/autogamess#2 @Andrew-S-Rosen
I have read through the paper once more and am satisfied with the corrections. I have a few remaining, minor comments on this part of the submission:

* "As well as packages for data collection from the output files of computational chemistry software have been made" is not a full a sentence
* "Specific packages have also been developed for to compliment". Delete "for"
* ", this requirement" --> ". This requirement"
* Although functionally equivalent, please have the `pip` installation instructions be identical between the paper and GitHub README
* It would be helpful to list the required dependencies on the GitHub README as well
* To aid readability, I would recommend splitting the large block of text into paragraphs. For instance, the second paragraph could start at "Oftentimes, single calculations" and a third starting at "AutoGAMESS provides an open-source"
