### 2019-08-05 ceff13a
Pip install now installs dependancies too fixes #3
