+++ setup.py
@@ -13,6 +13,8 @@ setuptools.setup(
     long_description_content_type="text/markdown",
     url="https://github.com/Cavenfish/autogamess",
     packages=setuptools.find_packages(),
+    install_requires=["numpy", "scipy", "pandas", "basis_set_exchange", "PeriodicElements",
+                      "openpyxl"],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
