### 2019-08-05 3c0810f
addressed all extra comments from issue 2
