+++ Paper/paper.md
@@ -23,7 +23,11 @@ bibliography: paper.bib
 The *ab initio* Quantum Chemistry software GAMESS(us)[@schmidt1993general] is capable of calculating a variety of molecular
 properties. One of the many popular uses of GAMESS(us) is by chemists, physicists, astro-chemists and astro-physicists
 [@burda2004theoretical; @pacifici2013high; @hickman2005dissociative; @bennett2014experimental] for predicting properties of
-volatile and unstable species that have not been experimentally characterized or quantized before. Applications of and research done using GAMESS(us) is not limited to uncharacterized species, it's also widely used in material characterization or material property prediction research. Research utilizing these types of *ab initio* calculations typically require calculations with multiple steps required to achieve each final result. For instance a Raman activity prediction first requires a geometry optimization and Hessian calculation be performed on the molecule, making automation extremely beneficial. Oftentimes single calculations of molecular properties are not reliable resulting in publications requiring several calculations, each implementing either a different level of theory or basis set, for each property be done on each molecule. This has brought about a demand for high throughput data calculation packages which automate these *ab initio* calculations[@Bhoorasingh; @MATHEW2017140; @Hjorth_Larsen_2017]. As well as packages for data collection from the output files of computational chemistry software have been made[@o2008cclib]. Specific packages have also been developed for to compliment GAMESS(us) [@schaftenaar2000molden; @schmidt2013webmo; @bode1998macmolplt; @allouche2011gabedit; @dermardirossian2005gdis]. However, because these programs are largely visualization and graphical programs, there is still a need for packages that automate GAMESS(us) Raman calculations. As it stands, the automation of Raman calculations is either not being done, or being implemented individually by each research group utilizing the GAMESS(us) software. This slows scientific progress down, and an automation software written in a language extremely simple and well adopted by scientists, such as Python, is an attractive solution to the problem. AutoGAMESS provides an open-source Python based software for automating conversion between optimization calculations to Hessian calculations and then to Raman calculations. It also offers automation of data collection from the output files, for quick tabular data readouts of each calculation. AutoGAMESS has currently been used for a study presented at the 30th Annual Conference on Computation Physics CCP2018, its proceedings is soon to be published. AutoGAMESS is also currently being used in multiple other computational chemistry projects, soon to be published by scientists at the University of Central Florida.
+volatile and unstable species that have not been experimentally characterized or quantized before. Applications of and research done using GAMESS(us) is not limited to uncharacterized species, it's also widely used in material characterization or material property prediction research. Research utilizing these types of *ab initio* calculations typically require calculations with multiple steps required to achieve each final result. For instance a Raman activity prediction first requires a geometry optimization and Hessian calculation be performed on the molecule, making automation extremely beneficial.
+
[…]
