### 2019-08-05 407949f
addressed all extra comments from issue 2
