### 2019-08-05 Cavenfish/autogamess#4 @Andrew-S-Rosen
More detailed description of test case
This is part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1612).

**Required for consideration in JOSS**
There is currently no description of what test cases are available, how to run them, and what the expected output is. I see functions to test the code in the `/tests` directory, but is there a script that runs each test automatically to confirm everything is working properly? It currently looks like it's just a list of individual functions.

**Optional suggestions**
Also, with a test code is provided, it would also be helpful for the user if there is some feedback as to what is going on when the test case is run (beyond it just crashing if an assertion fails). For instance, something printed to the screen saying "Running test case: building input" followed by "Success" or something to let the user know that test passed (and so on).

While not a requirement for acceptance, it would improve the code if the test cases were run automatically with each commit (e.g. as part of Travis CI). This may be worth thinking about in the future, especially if new features are to be added to AutoGAMESS.
