### 2019-08-05 Cavenfish/autogamess#4 @Cavenfish
I have a very basic description of how to run the tests in the CONTRIBUTE.md .  I can also add this to the README.md, and will add some descriptive outputs in the functions soon.
