### 2019-08-05 ea8de40
added tests command to readme fixes #4
