### 2019-08-07 c5f5209
used numpy all close to fix 1 off error in get_data test, and used ignore .gitignore in dircmp to fix error in new_project test, fixes #5
