### 2019-08-07 ccefc8d
corrected spelling mistakes in readme, added missing doc strings, fixes #6
