+++ autogamess/get_data.py
@@ -6,6 +6,41 @@ class DATA:
         return
 
 def get_data(filename):
+    """
+    This function collects data from GAMESS(us) log files.
+
+    Parameters
+    ----------
+    filename: string
+        This should be a string that points to the log file of any
+        GAMESS(us) calculation. (FULL DIRECTORY STRING REQUIRED)
+
+    Returns
+    -------
+    data: object
+        This is an object with all the data collected from the log file.
+        Below is a list of the attributes associated with `data` based on
+        each log file type. 
+
+        all   files: `cpu`, `time`
+        opt   files: `bond_lengths`, `bond_angles`
+        hes   files: `vib_freq`, `ir_inten`
+        raman files: `raman`
+        vscf  files: `vscf_freq`, vscf_ir
+
+    Notes
+    -------
+    This function is primarily intended for interal use by AutoGAMESS.
+
+    Example
+    -------
+    >>> import autogamess as ag
+    >>>
+    >>> filename = './AGv0-0-6_NH3_CCSD-T_CC6_opt.log'
+    >>>
+    >>> ag.get_data(filename)
+    >>>
+    """
     data    = DATA()
 
     if '_opt' in filename:
