+++ examples/compile_data.py
@@ -1,0 +1,7 @@
+import autogamess as ag
+
+logsdir = './done/'
+projdir = './AutoGAMESS COx Project/'
+
+ag.sort_logs(projdir, logsdir)
+ag.fill_spreadsheets(projdir)
