### 2019-08-07 66bdb19
Made them into template scripts
