+++ examples/compile_data.py
@@ -1,7 +1,7 @@
 import autogamess as ag
 
-logsdir = './done/'
-projdir = './AutoGAMESS COx Project/'
+logsdir = ''#Define here the logs directory
+projdir = ''#Define here the project directory
 
 ag.sort_logs(projdir, logsdir)
 ag.fill_spreadsheets(projdir)
