### 2019-08-07 66fd978
I have addressed all of the reviewers concerns, fixes #7
