+++ Paper/paper.bib
@@ -189,3 +189,13 @@
 	title={The atomic simulation environment{\textemdash}a Python library for working with atoms},
 	journal={Journal of Physics: Condensed Matter}
 }
+
+@article{KROGEL2016154,
+title={Nexus: A modular workflow management system for quantum simulation codes},
+journal={Computer Physics Communications},
+volume={198},
+pages={154 - 168},
+year={2016},
+doi={10.1016/j.cpc.2015.08.012},
+author={Jaron T. Krogel}
+}
