### 2019-08-08 Cavenfish/autogamess#9 @colleeneb
general misc comments for JOSS paper
This is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/1612). I’m only partly through the review, though.

First, thank you for this handy software!

General comments:

- In README.md ("This is a python module for automating Raman calculations using the GAMESS(us) Quantum Chemistry software.") and in the second paragraph of paper.md, it almost sounds like the software is about automating the running of GAMESS jobs. However, it seems that autoGAMESS only automates generation of input files/parsing of log files. (Please correct me if I’m wrong.) Automation of input file creation and log file parsing is extremely useful, but it would be helpful to be clear about the functionality in README.md.
- In README.md, please note somewhere which version of GAMESS this was developed using / which one it works with, since you are parsing a GAMESS log/input file, and things sometimes change. I doubt the log files will change much in structure, but it’s possible they might, so if you point out which one version you know it works with, it might prevent unexpected problems that people have.

In README.md under the “Examples” section:

- Possibly a typo in the naming convention—Species instead of specie?
- Could you give an example in addition to the the convention? There is one below in the “less common method of utlization” section, but it would be nice if it was moved above (it’s useful to have both naming convention and examples). So even just adding,

`An example is “AG-test_H2O_B3LYP_CCD_opt.log”, where the arbitrary_thing is AG-test, Species is H2O, Theory Level is…` would be nice.

- “Since AutoGAMESS reads information from file names and requires the underscore separates the information something must be present there to prevent confusion”. What happens when the name doesn’t match the convention? A crash, an error message, or just incorrect data read?
- There is a sentence saying “This will retrieve the data from the file, it will read the file name to get the run type.” Could you also describe what it’s doing in more detail? `get_data` is not described in the Function documentation section. I know you mention that it’s an internally used function, but if you’re going to give it as an example in the “Examples” section, it should be documented.

In paper.md:

- Grammatical: “AutoGAMESS has currently been used for a study presented at the 30th Annual Conference on Computation Physics CCP2018, its proceedings is soon to be published. “ is a little strange to my ears. I suggest: “AutoGAMESS has currently been used for a study presented at the 30th Annual Conference on Computation Physics CCP2018, and will be published in the conference proceedings. “
- Per https://www.msg.chem.iastate.edu/gamess/references.html, reference both "General Atomic and Molecular Electronic Structure System" M.W.Schmidt, K.K.Baldridge, J.A.Boatz, S.T.Elbert, M.S.Gordon, J.H.Jensen, S.Koseki, N.Matsunaga, K.A.Nguyen, S.Su, T.L.Windus, M.Dupuis, J.A.Montgomery J. Comput. Chem., 14, 1347-1363(1993).
And
"Advances in electronic structure theory: GAMESS a decade later" M.S.Gordon, M.W.Schmidt pp. 1167-1189, in "Theory and Applications of Computational Chemistry: the first forty years" C.E.Dykstra, G.Frenking, K.S.Kim, G.E.Scuseria (editors), Elsevier, Amsterdam, 2005.
