### 2019-08-08 b0473ea
Added second GAMESS(US) citation
