+++ Paper/paper.md
@@ -20,8 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-The *ab initio* Quantum Chemistry software GAMESS(US)[@schmidt1993general] is capable of calculating a variety of molecular
-properties. One of the many popular uses of GAMESS(us) is the prediction of properties of volatile and unstable species that have not been experimentally characterized or quantized before by chemists, physicists, astro-chemists and astro-physicists[@burda2004theoretical; @pacifici2013high; @hickman2005dissociative; @bennett2014experimental]. Applications of and research done using GAMESS(US) is not limited to uncharacterized species, it's also widely used in material characterization or material property prediction research. Research utilizing these types of *ab initio* calculations typically require calculations with multiple steps required to achieve each final result. For instance a Raman activity prediction first requires a geometry optimization and Hessian calculation be performed on the molecule, making automation extremely beneficial. This leads to complicated and tedious workflows slowing a user's research.
+The *ab initio* Quantum Chemistry software GAMESS(US)[@schmidt1993general; @GORDON20051167] is capable of calculating a variety of molecular properties. One of the many popular uses of GAMESS(US) is the prediction of properties of volatile and unstable species that have not been experimentally characterized or quantized before by chemists, physicists, astro-chemists and astro-physicists[@burda2004theoretical; @pacifici2013high; @hickman2005dissociative; @bennett2014experimental]. Applications of and research done using GAMESS(US) is not limited to uncharacterized species, it's also widely used in material characterization or material property prediction research. Research utilizing these types of *ab initio* calculations typically require calculations with multiple steps required to achieve each final result. For instance a Raman activity prediction first requires a geometry optimization and Hessian calculation be performed on the molecule, making automation extremely beneficial. This leads to complicated and tedious workflows slowing a user's research.
 
 Oftentimes single calculations of molecular properties are not reliable resulting in publications requiring several calculations, each implementing either a different level of theory or basis set, for each property be done on each molecule. This has brought about a demand for high throughput data calculation packages which automate these *ab initio* calculations[@Bhoorasingh; @MATHEW2017140; @Hjorth_Larsen_2017]; as well as workflow management systems [@KROGEL2016154]. Specific packages have also been developed to compliment GAMESS(US) [@schaftenaar2000molden; @schmidt2013webmo; @bode1998macmolplt; @allouche2011gabedit; @dermardirossian2005gdis]. However, because these programs are largely visualization and graphical programs, there is still a need for packages that automate GAMESS(US) Raman calculations. Automation is essential to generate large databases of Raman data, which could have further applications for machine learning of Raman data. As it stands, the automation of Raman calculations is either not being done, or being implemented individually by each research group utilizing the GAMESS(US) software. This slows scientific progress down, and an automation software written in a language extremely simple and well adopted by scientists, such as Python, is an attractive solution to the problem.
 
