+++ Paper/paper.bib
@@ -10,6 +10,17 @@
   publisher={Wiley Online Library}
 }
 
+@incollection{GORDON20051167,
+title={Chapter 41 - Advances in electronic structure theory: GAMESS a decade later},
+booktitle={Theory and Applications of Computational Chemistry},
+publisher={Elsevier},
+address={Amsterdam},
+pages={1167 - 1189},
+year={2005},
+doi={10.1016/B978-044451719-7/50084-6},
+author={Mark S. Gordon and Michael W. Schmidt}
+}
+
 @article{burda2004theoretical,
   title={Theoretical model of copper Cu (I)/Cu (II) hydration. DFT and ab initio quantum chemical study},
   author={Burda, Jaroslav V and Pavelka, Mat{\v{e}}j and {\v{S}}im{\'a}nek, Milan},
