### 2019-08-08 Cavenfish/autogamess#9 @Cavenfish
Hey @colleeneb the `Species` header is not a typo, it's assumed multiple specie will be put in the `input.csv` so i default the header to be plural.

As for the rest of your comments, I am currently fixing all issues.

Thanks!
