### 2019-08-08 8d3f63e
Added requirements text file, fixes #8
