### 2019-08-11 Cavenfish/autogamess#10 @colleeneb
question about input_builder 
This is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/1612).

I'm trying to test input_builder using the examples of the CSV files that are discussed in the README.

I set up input.csv to contain data based on the README:
```
$ cat input.csv
Species,Theory,Basis Sets,External Basis Sets,Run Types
H2O,B3LYP,CCD,may-cc-pVDZ,Optimization
```

and then I tried to call input_builder using it, and get an error saying that `Basis set may-cc-pv(d+d)z does not exist`. Should I use something other than `may-cc-pVDZ` in the CSV? (I copied it from the README initially.)

```
$ python
>>> import autogamess as ag
>>> csvfile = './input.csv'
>>> savedir= './'
>>> ag.input_builder(csvfile, savedir)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/colleen/autogamess/autogamess/input_builder.py", line 170, in input_builder
    fmt='gamess_us', header=False).split('\n')[1:]
  File "/Users/colleen/.local/lib/python3.6/site-packages/basis_set_exchange/api.py", line 168, in get_basis
    bs_data = _get_basis_metadata(name, data_dir)
  File "/Users/colleen/.local/lib/python3.6/site-packages/basis_set_exchange/api.py", line 53, in _get_basis_metadata
    raise KeyError("Basis set {} does not exist".format(name))
KeyError: 'Basis set may-cc-pv(d+d)z does not exist'
```

Am I doing something wrong? Thanks for any help!
