### 2019-08-11 Cavenfish/autogamess#10 @colleeneb
I'll add a second issue I ran into:

I can get the input file to be generated if I remove the external basis set, so that input.csv is:
```
bash-3.2$ cat input.csv
Species,Theory,Basis Sets,External Basis Sets,Run Types
H2O,B3LYP,CCD,,Optimization
```
However, when I try to use the `new_project` function with `make_inputs=True`, there is an error, since it looks like `new_project.py` has a hard-coded path to `inputs/`:
```
$ python
>>> import autogamess as ag
>>> csvfile = './input.csv'
>>> maindir= './'
>>> title   = 'Project Title/'
>>> ag.new_project(maindir, csvfile, title=title,make_inputs=True)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/colleen/autogamess/autogamess/new_project.py", line 166, in new_project
    proj_title = title.replace('/', '\n'))
  File "/Users/colleen/autogamess/autogamess/input_builder.py", line 97, in input_builder
    f = open(save_dir + filename, 'w')
FileNotFoundError: [Errno 2] No such file or directory: './inputs/AGv1-0-35_H2O_B3LYP_CCD_opt.inp'
>>>

```
