### 2019-08-12 Cavenfish/autogamess#10 @colleeneb
Third issue I ran into:

In "Function Documentations", under the "Examples" section in "sort_log", the example code doesn't seem to work.
For reference, the example code is:

```
Examples
----------
>>>import sort_logs
>>>projdir = './Example/'
>>>sort_logs(projdir)
>>>
```
Based on the function declaration, it needs two inputs, and `import sort_logs` doesn't seem to work.
