### 2019-08-12 Cavenfish/autogamess#10 @Cavenfish
Hey @colleeneb, first off thank you for the comments. In terms of the errors you have posted:

1) It seems EMSL no longer carries the 'may-cc-pv(d+d)z' basis set, which is why the error came up. They still have the other double zeta calendar basis sets though, so if you did instead 'jun-cc-pv(d+d)z' it should work. I will change the README so that it is not there anymore.

2) This is a bug, I must have forgotten to modify the `savedir` variable. I will fix this right away. Thank you again for catching this.

3) It seems as though when i modified this doc string i forgot to modify the example section. This is an old example that, as you noticed, won't work for the current AutoGAMESS version. I will fix this right away.
