### 2019-08-12 f04fa16
removed may-cc-pVDZ from dictionary since emsl no longer carries it
