### 2019-08-12 755ca38
made the may basis set triple zeta and fixed sort_logs docstring fixes #10
