### 2019-08-12 85188f3
fixed examples section of docstring
