### 2019-08-12 Cavenfish/autogamess#11 @colleeneb
Question about fill_spreadsheet and continuing JOSS review 
This is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/1612).

Almost done! :) Just a few things left.

### Editorial comment:
In “Examples of Common AutoGAMESS Utilization Methods” section in README.md, there are 3 examples. Could these be titled separately (giving each example a heading), for clarity?

### General comment:
Since the test inputs were all from the Windows version of GAMESS, I tested if the log parsing worked with the Linux version of GAMESS to see if it could parse the log files and if the inputs it generated ran properly. Everything worked properly, at least for water :) .

### Issue:
I have been trying all the functions in README.md.  My overall goal was to generate data for Optimization, Hessian, and Raman for the H2O input, and then use fill_spreadsheet to parse the data into the spreadsheet. I ran into an issue with fill_spreadsheet, though (described below).

First, I was able to generate an input with `make_input`, run it with GAMESS, use `opt2hes` to generate a hessian input, run it with GAMESS, use `hes2raman` to get a Raman input, and then run it. I then used `new_project` to generate a directory structure, and `sort_logs` to sort the logs into the Logs/Sorted subdirectory. After doing this, I end up with all the logs in the `Logs/Sorted/H2O/` subdirectory:

```
bash-3.2$ ls  Project\ Title/Logs/Sorted/H2O/
AGv1-0-35_H2O_B3LYP_CCD_hes.log
AGv1-0-35_H2O_B3LYP_CCD_opt.log
AGv1-0-35_H2O_B3LYP_CCD_raman.log
```

However, one thing has tripped me up—after doing what I described above, when I tried to run fill_spreadsheets, it failed with this error:
```
Traceback (most recent call last):
  File "fill.py", line 11, in <module>
    ag.fill_spreadsheets(sorteddir=sorteddir, sheetsdir=sheetsdir)
  File "", line 174, in fill_spreadsheets
    temp = df[ram]
KeyError: 'Raman'
```
It looks like the input.csv I used to generate the project structure only had one line, for H2O Optimization, and it looks like to pull out Raman and Hessian information, these need to be in input.csv, too. So I tried to regenerate the spreadsheets and directory structure with the following input.csv
```
bash-3.2$ cat input.csv
Species,Theory,Basis Sets,External Basis Sets,Run Types
H2O,B3LYP,CCD,,Raman
H2O,B3LYP,CCD,,Hessian
H2O,B3LYP,CCD,,Optimization
```
But this input.csv failed to generate a directory structure:
```
bash-3.2$ python
>>> import autogamess as ag
>>> csvfile = './input.csv'
>>> maindir= './'
>>> title   = 'Project Title/'
>>> ag.new_project(maindir, csvfile, title=title)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/colleen/autogamess/autogamess/new_project.py", line 115, in new_project
    os.makedirs(inputs+runtyp+specie)
  File "/anaconda3/lib/python3.6/os.py", line 220, in makedirs
    mkdir(name, mode)
FileExistsError: [Errno 17] File exists: './Project Title/Inps/Raman/H2O/'
```
Any advice on how to proceed? Or maybe I missed something? My use case was to generate the data for Optimization, Hessian, and Raman, for H2O, and then just use fill_spreadsheet to parse the data into the spreadsheet. Maybe this is not a normal use case?
