### 2019-08-12 Cavenfish/autogamess#11 @Cavenfish
Hey @colleeneb, glad to hear the data parsing worked for Linux! I never tested it with the Linux version. I will make the headings in the README right away.

In terms of the issue you ran into, the CSV file should only have each item on there once, if you take a look at the input.csv in the tests directory it has the input for running water all the way to raman with B3LYP CCD.
It should look like this:

| Species | Theory | Basis Sets | External Basis Sets | Run Types |
| ------- | ------ | ---------- | ------------------- | --------- |
H2O | B3LYP | CCD| | Optimization
 | | | | | Hessian
 | | | |  | Raman

AutoGAMESS will generate a project directory assuming every combination of, basis sets (internal and external), theory and run type will be performed for all species. So in the above example it assumes H2O B3LYP CCD will run Optimization, Hessian and Raman.
