### 2019-08-12 651cadd
Added headers to examples section, fixes #11
