### 2019-08-12 Cavenfish/autogamess#11 @colleeneb
Thanks for the help. Could you mention somewhere in README.md something along the lines of what you said—“the CSV file should only have each item on there once, but could have multiple run types”, and add a few lines in the CSV example in the README about it? Just so other people don’t get tripped up.

Once I changed input.csv as suggested, the spreadsheet was generated as expected with `ag.fill_spreadsheets('./Project Title/')`.
However, it moved the log files from `Sorted` directory to `[project]/Pass/[run_type]/[species]/`, which I didn’t expect—could this be documented? I only expected it to fill the spreadsheets, not move files, too.

On that note, trying to run
```
import autogamess as ag
sorteddir = './Project Title/Logs/Sorted/'
sheetsdir = './Project Title/Spreadsheets/'
ag.fill_spreadsheets(sorteddir=sorteddir, sheetsdir=sheetsdir)
```
Using the same setup as before—log files for H20 for Optimization, Hessian, and Raman in “./Project Title/Logs/Sorted/“, running the code above resulted in the following error:
```
Traceback (most recent call last):
  File "fill.py", line 11, in <module>
    ag.fill_spreadsheets(sorteddir=sorteddir, sheetsdir=sheetsdir)
  File "/Users/colleen/autogamess/autogamess/fill_spreadsheets.py", line 203, in fill_spreadsheets
    move2 = passdir + ram + '/' + dir + '/' + file
UnboundLocalError: local variable 'passdir' referenced before assignment
```
It looks like in fill_spreadsheets.py, passedir is only defined if `projdir` is True, but it’s used later even if it's not true.
