### 2019-08-12 Cavenfish/autogamess#11 @Cavenfish
Hey @colleeneb I will add that to the README right away! I will also add that `fill_spreadsheets` moves the files to the documentation as well. I'm also going to fix the `passdir` issue, right away.

Again, as always thank you so much for the comments and bug spotting!
