### 2019-08-12 Cavenfish/autogamess#11 @colleeneb
(One possible suggestion is to put any example that is in the README.md also in the testing setup, so you know those are being tested. However, this is not required for the review/acceptance, just a general comment.)
