### 2019-08-12 83000d1
Added more explanation of how input_builder and new_project csvfile parameter should be formatted
