### 2019-08-12 bdc227a
Add note about how fill_spreadsheets will move log files after filling spreadsheet
