### 2019-08-12 edabe24
fixed `passdir` variable problem
