### 2019-08-12 c2a0cf6
fixed inline math
