### 2019-08-12 fefe43e
fixed inline math
