### 2019-08-13 ec52d2c
fixed inline math
