### 2019-08-13 Cavenfish/autogamess#11 @colleeneb
Last thing! (I'm staying in this issue since it's related. I repulled and tried running

```
import autogamess as ag
sorteddir = './Project Title/Logs/Sorted/'
sheetsdir = './Project Title/Spreadsheets/'
ag.fill_spreadsheets(sorteddir=sorteddir, sheetsdir=sheetsdir)
```

again, and now I get the error below. So passdir or something else is missing a `/` (it seems to be looking for `'./Project TitleLogs/Pass/Raman`.)

```
Traceback (most recent call last):
  File "fill.py", line 9, in <module>
    ag.fill_spreadsheets(sorteddir=sorteddir, sheetsdir=sheetsdir)
  File "/Users/colleen/autogamess/autogamess/fill_spreadsheets.py", line 213, in fill_spreadsheets
    os.rename(filename, move2)
FileNotFoundError: [Errno 2] No such file or directory: './Project Title/Logs/Sorted/H2O/AGv1-0-35_H2O_B3LYP_CCD_raman.log' -> './Project TitleLogs/Pass/Raman/H2O/AGv1-0-35_H2O_B3LYP_CCD_raman.log'
```
