### 2019-08-13 74e9209
projdir was missing `/` from previous fix
