### 2019-08-13 Cavenfish/autogamess#11 @Cavenfish
@colleeneb I just fixed it (at least I think i did lol). Let me know if its working now.
