### 2019-08-13 Cavenfish/autogamess#11 @colleeneb
Yes, that worked for me, thanks!
