### 2019-08-13 #1612 @colleeneb
@Cavenfish This is outside of the scope of the review, but I wanted to mention that if you are interested in making changes/improvements to GAMESS itself, you can follow the instructions [here](https://www.msg.chem.iastate.edu/gamess/developers.html), and email to request access to the GAMESS private repo on github.
