### 2019-08-13 Cavenfish/autogamess#11 @colleeneb
@Cavenfish This is outside of the scope of the review, but I wanted to mention that if you are interested in making changes/improvements to GAMESS itself, you can follow the instructions [here](https://www.msg.chem.iastate.edu/gamess/developers.html), and email to request access to the GAMESS private repo on github.
(I moved this from the review page to here, since I think it's more appropriate here than in the review page. Sorry for spamming you twice.)
