### 2019-08-13 Cavenfish/autogamess#11 @Cavenfish
@colleeneb Not a problem, twice isn't too bad! I will look into it, thank you for passing along this information.
