### 2019-08-13 #1612 @jedbrown
@Cavenfish Did you mean to cite cclib in the paper? I see it in paper.bib, but not cited in the paper.
