### 2019-08-13 #1612 @Cavenfish
@jedbrown  Yes, i did. It must have gotten dropped by accident during one of my edits to the paper. I'll add it back in now. I also want to mention before the publication I'd like to make a final version change, to 1.1.0 to make the version changes from publication onward easier to spot for users. I'm waiting for @arosen93  to finish with his review before making the version change though.
