### 2019-08-13 3b338be
quantized-> quantified, added cclib citation that got lost
