### 2019-08-14 Cavenfish/autogamess#13 @Andrew-S-Rosen
Graceful handling of existing folders
This is part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1612). `ag.new_project(...)` crashes when the directory is already present. I see a few ways this could be improved. Either you could create a new directory (e.g. append a number to the end of the user-provided name, like 'Water_01') or have the code print a warning to the user and gracefully close (instead of crashing with a FileExistsError). You can test this error out for yourself by running Step 5 of the Jupyter Notebook with the files and such already there.
