### 2019-08-14 Cavenfish/autogamess#2 @Andrew-S-Rosen
Minor remaining comments:
- "Research utilizing these... typically require" (require --> requires)
- Delete the second "required" in the above sentence
- Add a comma after "For instance"
- Add a comma after "are not reliable"
- "for each property be done" is not grammatically correct
- delete the semicolon before "as well as workflow"
- delete the comma before "for quick tabular"
- make "While bond angles" not capitalized
