### 2019-08-14 Cavenfish/autogamess#15 @Andrew-S-Rosen
Error in pytest
This is part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/1612).

Using the most recent version of AutoGAMESS, the automated test does not pass successfully. Please see the traceback below. In addition to the error in `test_input_builder`, I also point out the `openpyxl` deprecation warning, as that may come back to bite you in the butt sometime in the future.

```
======================================= test session starts ======================================== platform win32 -- Python 3.7.3, pytest-5.0.1, py-1.8.0, pluggy-0.12.0
rootdir: C:\Users\asros\Desktop\zzz\autogamess-master
plugins: arraydiff-0.3, doctestplus-0.3.0, openfiles-0.3.2, remotedata-0.3.1
collected 12 items

test_imports.py .......                                                                       [ 58%] test_package.py .F...                                                                         [100%]

============================================= FAILURES ============================================= ________________________________________ test_input_builder ________________________________________

tmpdir = local('C:\\Users\\asros\\AppData\\Local\\Temp\\pytest-of-asros\\pytest-1\\test_input_builder0')

    def test_input_builder(tmpdir):
        ag.input_builder(csvfile, tmpdir.strpath+'/')

        a = tmpdir.strpath+'/' + os.listdir(tmpdir.strpath+'/')[0]
        b = './correct/IBtest/' + os.listdir('./correct/IBtest/')[0]

        tmp=ag.read_file(a)
        tmp[0] = '! AutoGAMESS TEST\n'
        f=open(a,'w')
        f.writelines(tmp)
        f.close()

        if not fc.cmp(a,b):
>           assert 0
E           assert 0

test_package.py:33: AssertionError
========================================= warnings summary ========================================= C:\Users\asros\Anaconda3\lib\site-packages\openpyxl\xml\functions.py:31
  C:\Users\asros\Anaconda3\lib\site-packages\openpyxl\xml\functions.py:31: DeprecationWarning: defusedxml.lxml is no longer supported and will be removed in a future release.
    from defusedxml.lxml import fromstring as _fromstring, tostring

-- Docs: https://docs.pytest.org/en/latest/warnings.html
========================= 1 failed, 11 passed, 1 warnings in 3.51 seconds ==========================
```
