### 2019-08-14 #1612 @arosen93
Hi @Cavenfish. I've complete the testing of the code. Generally, it works as expected, especially as a result of the useful additions made over the course of the review process. I really like the improved examples as well.

@jedbrown: Once the pytest is fixed (https://github.com/Cavenfish/autogamess/issues/15), my review will be complete.
