### 2019-08-14 #1612 @jedbrown
Thanks, all.

@Cavenfish, I would like to relay a concern and suggestion raised in private discussion. There was some concern that the software as written will be thought of more as a utility that some users may not feel compelled to cite. It was proposed that adding the ability to plot results would aid reproducibility. For example, I understand that it's common to use GUI software such as MacMolPlt to make line plots of frequency versus infrared or Raman intensities. What do you think of teaching AutoGAMESS to create such plots (suitably named/labeled/annotated)?

Further (and perhaps out of scope at this time), is anyone here aware of an API or bulk download option for [CCCBDB](https://cccbdb.nist.gov/) or other repository of experimental data, such that software like AutoGAMESS could automate comparison to experiments?
