### 2019-08-14 Cavenfish/autogamess#15 @Cavenfish
Hey @arosen93  thank you for pointing out the deprecation warning, I'll make sure AutoGAMESS won't be hurt by it.

In terms of your error, I have found a solution and will be pushing it soon.
