### 2019-08-14 d2b3c91
got rid of fc.cmp and just doing list comparison now, fixes #15
