### 2019-08-14 #1612 @Cavenfish
Hey @jedbrown

In terms of AutoGAMESS being cited, I understand the concerns but I have spoken with other researchers at my University and they consider AutoGAMESS to be worth citing (and they plan on doing so when able). Of course, this is special circumstance and is not to be expected for all users of AutoGAMESS. If you think adding plots will increase the chance of citation then I will add the plotting feature to AutoGAMESS, I have considered adding something like that before but figured users would prefer to simply generate their own plots instead.

In terms of the CCCBDB comparison to experiments, this was actually something I tried to do at first and even found this [package](https://github.com/marcelo-mason/cccbdb-calculation-parser). However, it was not exactly what I was looking for, so I dropped it. I would be very interested in something that works better so that I could add a feature for comparing results as well as generating scaling factors.
