### 2019-08-14 Cavenfish/autogamess#13 @Cavenfish
@arosen93  That's a great suggestion, ill add that in right away!
