### 2019-08-14 88744d9
added a catch for directory already exists error, fixes #13
