### 2019-08-14 d6d8597
typo on error message corrected
