### 2019-08-14 #1612 @jedbrown
I think plotting would add significant value and improve workflow reproducibility. Thanks for your willingness to add this feature.

What was the deficiency in that CCBDB retriever/parser relative to what you were looking for? Are any reviewers aware of better tools for this?
