### 2019-08-14 #1612 @Cavenfish
The parser seems to only pull calculated data not experimental data from what I could tell. What I would like is the experimental data, which i considered leaving a feature request for but the code has not been changed in 2 years so I figured its no longer maintained.
