### 2019-08-14 #1612 @Cavenfish
My thoughts were to just remake the package entirely from scratch.
