### 2019-08-14 3a716ee
added make_plot function import
