### 2019-08-15 9b51373
added documentation for make_plot
