### 2019-08-15 f082f6e
added stuff about make_plot function
