### 2019-08-15 51a593c
added matplotlib to dependancies
