+++ setup.py
@@ -14,7 +14,7 @@ setuptools.setup(
     url="https://github.com/Cavenfish/autogamess",
     packages=setuptools.find_packages(),
     install_requires=["numpy", "scipy", "pandas", "basis_set_exchange", "PeriodicElements",
-                      "openpyxl"],
+                      "openpyxl", "matplotlib"],
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
