+++ requirements.txt
@@ -4,3 +4,4 @@ pandas
 basis_set_exchange
 PeriodicElements
 openpyxl
+matplotlib
