### 2019-08-15 157c44a
added example of line plots
