+++ Paper/paper.md
@@ -52,9 +52,7 @@ where $A$ is the bond angle, $\hat{V_1}$ and $\hat{V_2}$ are the normalized posi
 
 AutoGAMESS is also capable of generating line plots of the vibrational frequency vs. IR/Raman intensity. Generated plots will be titled with the molecule name in the file and the theory/basis set used for the calculation. Each symmetry group will be plotted in a different color, from either a default or user specified color list. Figure 1 below shows an example line plot of H2O using B3LYP/CCD for the calculation.
 
- IR Line Plot | Raman Line Plot
- -------------|----------------
-![](AG-test_H2O_B3LYP_CCD_ir-plot.png) | ![](AG-test_H2O_B3LYP_CCD_raman-plot.png)
+![Line Plots Example](line-plots.png)
 
 # Use of AutoGAMESS
 
