### 2019-08-15 #1612 @Cavenfish
I apologize for the whedon spam, the figure gave me some trouble.

@jedbrown  I have added the plotting feature to AutoGAMESS, an example is now included in both the README and paper.

If any of you have any feedback about the plots please let me know.
