### 2019-08-16 #1612 @shivupa
I have some points that I think would strengthen this submission. These are open to discussion There has been mention that the scope is somewhat limited. I think adopting some of the points below would help broaden the scope and increase adoption in the target audience.

Short term:

1. Presenting the examples as Jupyter notebook is helpful, but adding Binder hosting would really help. Some of the users could be experimentalists who aren't used to the command line which may limit adoption otherwise. The notebook currently in examples would need to have the bash cells commented out as the binder instances would not have GAMESS in them.

2. I think the addition of plotting was very necessary. Something I would like to see (aka would make me use this over just making the plot myself) is line broadening for visual assistance with choices for line shapes (Gaussian/Lorentzian) and other associated parameters.

3. Integration with cclib would be nice. Short term, I would like to see at least the initial molecule being read with cclib and then being fed into the new project function. Clearly the parsing of the necessary data isn't being done yet in cclib, but either implementing this upstream or opening issues for these would allow you to support many codes.

4. I'd really like to see the addition of the generation of scaling factors. For now instead of interfacing with a database, I would like to see some representation of an experimental spectrum being read in and being matched to calculated intensities.

General point / Long term:

5. I think this should be called Autovib or AutoRAMAN because it automates these types of calculations rather than the full spectrum of calculations that GAMESS is capable of. Rebranding it as such would allow for (future work) adding other software packages such as Orca, PYSCF, etc.

Let me know what you think.
