### 2019-08-16 #1612 @Cavenfish
Hey @shivupa you make some great points,

1) I will need to learn more about Binder before attempting to implement it. This is the same with Sphinx which has also been suggested.

2) I can have that feature added right away (it is also something my adivsor recommended), my thought is to have the line broadening choices be over-plotted with a different line-style (dotted or dashed)  so that the vertical lines remain but you also have the broadened lines in the plot as well. What do you think, is that too much for one plot?

3) I can replace the default initial molecule library with a cclib interface. I just don't see how to make that happen, does cclib have a pre-optimization molecule coordinate dictionary or some way of getting me the point group symmetry and initial atom coordinates? If so, can you please help me find it.

4) I can make a function that has users give experimental data in list form and it then generates the scaling factors (as well as the scaled frequencies) from there.

5) Re-branding it as AutoRaman seems like it may be an inconvenience but I do see merit in that suggestion. I will consider it.
