### 2019-08-16 b2b6bb0
added line broadening and corrected typo fixes #16
