+++ autogamess/make_plot.py
@@ -2,8 +2,10 @@ from .config import *
 from .get_data import get_data
 from .data_finder import hessian
 import matplotlib.pyplot as plt
+from scipy.optimize import curve_fit
 
-def make_plot(file, savedir, cmap=['b', 'k', 'r', 'g', 'y', 'c']):
+def make_plot(file, savedir, cmap=['b', 'k', 'r', 'g', 'y', 'c'],
+              method='None', sig=300):
     """
     This function make vibrational frequency vs. IR/Raman intensity line plots.
 
@@ -18,6 +20,9 @@ def make_plot(file, savedir, cmap=['b', 'k', 'r', 'g', 'y', 'c']):
     cmap: list [Optional]
         This should be a list of Matplotlib allowed color choices. Each symmetry
         will be plotted with a different color in the list.
+    method: string [Optional]
+        This should be string giving the method for line broadening, options are
+        `Gaussian`, `Lorentzian`, 'None'(defualt).
 
     Returns
     -------
@@ -50,13 +55,40 @@ def make_plot(file, savedir, cmap=['b', 'k', 'r', 'g', 'y', 'c']):
         #init plot plane
         fig, ax = plt.subplots()
 
+        #get x max
+        x_max = 0
+        for key in vf:
+            x = np.float_(vf[key])
+            if max(x) > x_max:
+                x_max = max(x)
+
         #iterate through data making lines
-        i = 0
+        i    = 0
+        x2   = np.arange(0, x_max*1.25, 0.01)
+        sum  = np.zeros(len(x2))
         for key in vf:
             x = np.float_(vf[key])
             y = np.float_(ir[key])
             ax.vlines(x, 0, y, label=key, colors=cmap[i])
+
+            #Make Gaussian line broadening
+            if method is 'Gaussian':
+                for a,b in zip(x,y):
+                    gfit = gaussian(x2, a, sig, b)
+                    sum += gfit
+                    ax.plot(x2, gfit, linestyle='--', color=cmap[i])
+
+            #Make Lorentzian line broadening
+            if method is 'Lorentzian':
+                for a,b in zip(x,y):
+                    lfit = lorentzian(x2, a, sig, b)
+                    sum += lfit
+                    ax.plot(x2, lfit, linestyle='--', color=cmap[i])
+
             i += 1
+
+        #plot fit sum
+        ax.plot(x2, sum, alpha=0.5, color='r', label='Spectral Line')
 
         #init strings
         a     = file.split('_')
@@ -69,7 +101,7 @@ def make_plot(file, savedir, cmap=['b', 'k', 'r', 'g', 'y', 'c']):
         plt.ylim(bottom=0)
         plt.legend(loc='upper left')
         plt.xlabel(r'Vibrational Frequency $(cm^{-1})$')
-        plt.ylabel(r'Infrared Intensity $(Debey^2 \cdot \AA^{-2} \cdot amu^{-\frac{1}{2}})$')
+        plt.ylabel(r'Infrared Intensity $(Debye^2 \cdot \AA^{-2} \cdot amu^{-\frac{1}{2}})$')
         plt.title(title, fontsize=20)
         plt.savefig(savedir + name + png)
 
@@ -81,13 +113,40 @@ def make_plot(file, savedir, cmap=['b', 'k', 'r', 'g', 'y', 'c']):
         #init plot plane
         fig, ax = plt.subplots()
 
+        #get x max
+        x_max = 0
+        for key in vf:
+            x = np.float_(vf[key])
+            if max(x) > x_max:
+                x_max = max(x)
+
         #iterate through data making lines
         i = 0
+        x2   = np.arange(0, x_max*1.25, 0.01)
+        sum  = np.zeros(len(x2))
         for key in vf:
             x = np.float_(vf[key])
[…]
