### 2019-08-16 bc09a4b
added gaussian and lorentzian functions
