### 2019-08-16 4fd02d4
added more explanation about new feature with make_plot
