### 2019-08-16 #1612 @arosen93
Hey @Cavenfish. I think these are some good ideas. With regards to #2, I think what you said would be a great choice. You could also have an optional flag to hide the vertical lines if desired, but I think what you suggested would be a good default. Lots of visualization programs have a similar thing (I personally like ChemCraft's -- that might be a good reference). With regards to #5, I had the same thought as @shivupa, although didn't bring it up. I do think the name "AutoGAMESS" is a bit misleading, given the specific nature of the program, but I also understand that it may be inconvenient to "re-brand".
