### 2019-08-17 #1612 @Cavenfish
@shivupa @arosen93 I have added the line broadening feature, an example is in the README. I also added a spectral line plot which is the sum of the fits.
