### 2019-08-17 579ce5a
updated the make_plots explanation
