### 2019-08-17 da7c23e
updated doc string
