### 2019-08-17 8e8c7be
updated documentation for make_plot
