### 2019-08-17 d839f06
added scaling factor function import
