### 2019-08-17 f657643
added Scott least square method of scaling factor calculation
