### 2019-08-17 accc63e
added scaling factor function documentation
