### 2019-08-18 #1612 @shivupa
I opened an issue which shows how to address 3. I realized the library you would want to use is openbabel not cclib. I included an example of how to do this. It seems like code for 4 is added once we can review that and see an example I would say this address all of the issues that I think are essential (1 and 5 are long term and optional).
