### 2019-08-18 5185b54
added flag for how much to plot, added tight layout, added saveless option. fixes #17
