### 2019-08-18 5fe387a
removed commented out code
