### 2019-08-18 1904f65
updated make_plot documentation
