### 2019-08-18 e3c59e5
made it so spectral line is not plotted if line broadening is not done
