### 2019-08-18 516b84f
added imaginary vline plot when vline flagged off just to keep label in legend
