+++ autogamess/make_plot.py
@@ -98,6 +98,8 @@ def make_plot(file, savedir=None, cmap=['b', 'k', 'r', 'g', 'y', 'c'],
             #check flag, plot v lines
             if 1 not in flag:
                 ax.vlines(x, 0, y, label=key, colors=cmap[i])
+            else:
+                ax.vlines(0, 0, 0, label=key, colors=cmap[i])
 
             #Make Gaussian line broadening
             if method is 'Gaussian':
