### 2019-08-18 05e37a7
improved conditional for plotting that imaginary thing
