### 2019-08-18 70cc7db
improved conditional for plotting that imaginary thing
