### 2019-08-18 04e2844
updated documentation about initial_coords_dict, fixes #18
