### 2019-08-18 ffe38d6
added examples for the scaling factors and plotting fucntions
