### 2019-08-18 fb0969a
added scaling factor citation and explanation in paper capabilities section
