+++ Paper/paper.bib
@@ -210,3 +210,14 @@ year={2016},
 doi={10.1016/j.cpc.2015.08.012},
 author={Jaron T. Krogel}
 }
+
+@article{Scott_Radom,
+author = {Scott, Anthony P. and Radom, Leo},
+title = {Harmonic Vibrational Frequencies:  An Evaluation of Hartree−Fock, Møller−Plesset, Quadratic Configuration Interaction, Density Functional Theory, and Semiempirical Scale Factors},
+journal = {The Journal of Physical Chemistry},
+volume = {100},
+number = {41},
+pages = {16502-16513},
+year = {1996},
+doi = {10.1021/jp960976r}
+}
