### 2019-08-18 #1612 @Cavenfish
I have finished both the plotting function and scaling factor generation function. I have added documentation and examples for them in the README and put more in the paper about these two new capabilities.
