### 2019-08-18 5cff3e4
got rid of straggler phrase
