### 2019-08-18 #1612 @Cavenfish
i just pushed a commit that gets rid of the straggler phrase at the end of the paper's capabilities section.
