### 2019-08-21 #1612 @Cavenfish
@jedbrown @shivupa @arosen93 @colleeneb

I believe all the new features should be fully functional and reflected properly in the README and JOSS paper. I await feedback from you all, thanks again for the time and effort.
