### 2019-08-21 #1612 @arosen93
@jedbrown, @Cavenfish This is an excellent update to AutoGAMESS that I think really strengthens the use-cases for the code. I also like the ability to generate the scaling factors. The only remaining suggestion I have is to add a formal test case that generates an example IR and/or Raman plot (just like you have in the README) to confirm that this part of the code doesn't break. As long as the code doesn't crash when it makes the figure, I think that's a suitable test (it's easy for a `matplotlib` update to change something unexpected in one's code). Otherwise, I do not have any additional comments at this time.
