### 2019-08-21 #1612 @colleeneb
I also think this strengthens the use-cases for AutoGAMESS. It was a lot of nice work!  I have no additional comments either.
