### 2019-08-22 5fbee44
added make_plot test
