### 2019-08-22 #1612 @Cavenfish
@jedbrown @shivupa @arosen93 @colleeneb

I added the test for `make_plot` function to the package tests. I have also updated and released version 1.1.0 of AutoGAMESS, which includes all the new features.
