### 2019-08-24 #1612 @jedbrown
@Cavenfish Thanks for your great additions. I just returned from travel and the semester is about to start, but I'll handle this within the coming days. Sorry for the delay.
