### 2019-09-02 66fef98
Paper: use tt to distinguish function names
