### 2019-09-02 09edad9
Paper: fix citation spacing/style, general copy editing
