### 2019-09-02 897512c
Paper: bib formatting/punctuation
