+++ Paper/paper.bib
@@ -1,7 +1,7 @@
 @article{schmidt1993general,
   title={General atomic and molecular electronic structure system},
   author={Schmidt, Michael W and Baldridge, Kim K and Boatz, Jerry A and Elbert, Steven T and Gordon, Mark S and Jensen, Jan H and Koseki, Shiro and Matsunaga, Nikita and Nguyen, Kiet A and Su, Shujun and others},
-  journal={Journal of computational chemistry},
+  journal={Journal of Computational Chemistry},
   volume={14},
   number={11},
   doi={10.1002/jcc.540141112},
@@ -22,7 +22,7 @@ author={Mark S. Gordon and Michael W. Schmidt}
 }
 
 @article{burda2004theoretical,
-  title={Theoretical model of copper Cu (I)/Cu (II) hydration. DFT and ab initio quantum chemical study},
+  title={Theoretical model of copper {Cu(I)/Cu(II)} hydration. DFT and ab initio quantum chemical study},
   author={Burda, Jaroslav V and Pavelka, Mat{\v{e}}j and {\v{S}}im{\'a}nek, Milan},
   journal={Journal of Molecular Structure: THEOCHEM},
   volume={683},
@@ -34,9 +34,9 @@ author={Mark S. Gordon and Michael W. Schmidt}
 }
 
 @article{pacifici2013high,
-  title={A high-level ab initio study of the N2+ N2 reaction channel},
+  title={A high-level ab initio study of the $\mathrm N_2$ + $\mathrm N_2$ reaction channel},
   author={Pacifici, Leonardo and Verdicchio, Marco and Lago, Noelia Faginas and Lombardi, Andrea and Costantini, Alessandro},
-  journal={Journal of computational chemistry},
+  journal={Journal of Computational Chemistry},
   volume={34},
   number={31},
   doi={10.1002/jcc.23415},
@@ -46,7 +46,7 @@ author={Mark S. Gordon and Michael W. Schmidt}
 }
 
 @article{hickman2005dissociative,
-  title={Dissociative recombination of e+ HCNH+: Diabatic potential curves and dynamics calculations},
+  title={Dissociative recombination of e + $\mathrm{HCNH}^+$: Diabatic potential curves and dynamics calculations},
   author={Hickman, AP and Miles, RD and Hayden, C and Talbi, D},
   journal={Astronomy \& Astrophysics},
   volume={438},
@@ -58,7 +58,7 @@ author={Mark S. Gordon and Michael W. Schmidt}
 }
 
 @article{bennett2014experimental,
-  title={Experimental studies on the formation of D2O and D2O2 by implantation of energetic D+ ions into oxygen ices},
+  title={Experimental studies on the formation of $\mathrm D_2\mathrm O$ and $\mathrm D_2\mathrm O_2$ by implantation of energetic $\mathrm D^+$ ions into oxygen ices},
   author={Bennett, Chris J and Ennis, Courtney P and Kaiser, Ralf I},
   journal={The Astrophysical Journal},
   volume={782},
@@ -72,7 +72,7 @@ author={Mark S. Gordon and Michael W. Schmidt}
 @article{schaftenaar2000molden,
   title={Molden: a pre-and post-processing program for molecular and electronic structures},
   author={Schaftenaar, Gijs and Noordik, Jan H},
-  journal={Journal of computer-aided molecular design},
+  journal={Journal of Computer-Aided Molecular Design},
   volume={14},
   number={2},
   doi={10.1023/A:1008193805436},
@@ -90,7 +90,7 @@ author={Mark S. Gordon and Michael W. Schmidt}
 }
 
 @article{bode1998macmolplt,
-  title={Macmolplt: a graphical user interface for GAMESS},
[…]
