### 2019-09-02 #1612 @jedbrown
@Cavenfish Thanks for your patience. In addition to the PR, please consider whether your use of packages such as [SciPy](https://www.scipy.org/citing.html) and [Matplotlib](https://matplotlib.org/3.1.1/citing.html) (and others) warrant citation. Most likely yes.

The PyPI link (https://pypi.org/project/autogamess/) says
> The author of this package has not provided a project description

If you intend for that page to be the landing page for those who wish to install AutoGAMESS, then you'll want to curate its content. I would recommend cutting that link, such that a prospective user arrives at your GitHub and/or just installs via `pip`, typically without visiting the PyPI site. Also note that the repository is always linked from the sidebar in the typeset version so you don't need to link it again.

The wording in the section "Use of AutoGAMESS" is rough, but why rule out managing an ensemble using only Python (versus shell scripts calling user-written Python scripts calling AutoGAMESS)? Perhaps just say that you provide examples of incorporating AutoGAMESS into workflows based on shell scripts?
