### 2019-09-03 b660d99
added package citations, and edited wording form two sections
