### 2019-09-03 b0470a8
corrected package citations
