+++ Paper/paper.bib
@@ -230,6 +230,17 @@ doi = {10.1021/jp960976r}
   publisher={Trelgol Publishing USA}
 }
 
+@article{numpy,
+author = {Walt,Stéfan van der  and Colbert,S. Chris  and Varoquaux,Gaël },
+title = {The NumPy Array: A Structure for Efficient Numerical Computation},
+journal = {Computing in Science \& Engineering},
+volume = {13},
+number = {2},
+pages = {22-30},
+year = {2011},
+doi = {10.1109/MCSE.2011.37}
+}
+
 @Misc{scipy,
   author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   title =     {{SciPy}: Open source scientific tools for {Python}},
