### 2019-09-03 a970af6
added numpy citation
