### 2019-09-03 #1612 @Cavenfish
@jedbrown Hey, I merged your commits and made the changes you suggested above. Some of the package citations I wasn't able to find DOI's for, is this an issue?
