### 2019-09-03 #1612 @jedbrown
Another small PR, then I think we'll be ready. We only need DOIs for references that have them.
