### 2019-09-03 #1612 @jedbrown
Thanks! Looks good to me now, so please tag a release ([annotated tags](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) preferred). Presumably this will be v1.1.1, but it's your choice. Then please archive on Zenodo or a similar service and report the DOI back in this thread.
