### 2019-09-03 #1612 @Cavenfish
v1.1.1 sounds good to me! I made the tag and here is the DOI:  10.5281/zenodo.3385177.
