### 2019-09-03 #1612 @jedbrown
Thanks.

Sorry, my instructions may have been unclear: You should archive your _repository_ with Zenodo, not the paper. JOSS handles the paper DOI, with reference to the DOI for your repository.
