### 2019-09-03 #1612 @jedbrown
DOI are intended to be permanent so either leave it or add a new version (https://help.zenodo.org/#versioning).
