### 2019-09-03 #1612 @jedbrown
@openjournals/joss-eics This paper is accepted; over to you.
